+++ lib/maze/capabilities.rb
@@ -4,41 +4,6 @@ module Maze
   # Appium capabilities for each target farm
   class Capabilities
     class << self
-      # @param device_type [String] A key from @see BrowserStackDevices::DEVICE_HASH
-      # @param local_id [String] unique key for the tunnel instance
-      # @param capabilities_option [String] extra capabilities provided on the command line
-      def for_browser_stack_device(device_type, local_id, appium_version, capabilities_option)
-        capabilities = {
-          'bstack:options' => {
-            'local' => 'true',
-            'localIdentifier' => local_id
-          },
-          'noReset' => 'true'
-        }
-        capabilities.deep_merge! BrowserStackDevices::DEVICE_HASH[device_type]
-        capabilities.deep_merge! JSON.parse(capabilities_option)
-        capabilities['bstack:options']['appiumVersion'] = appium_version unless appium_version.nil?
-        capabilities
-      end
-
-      # @param browser_type [String] A key from @see browsers_bs.yml
-      # @param local_id [String] unique key for the tunnel instance
-      # @param capabilities_option [String] extra capabilities provided on the command line
-      def for_browser_stack_browser(browser_type, local_id, capabilities_option)
-        capabilities = {
-          'bstack:options' => {
-            'local' => 'true',
-            'localIdentifier' => local_id,
-            "os" => "Windows",
-            "osVersion" => "8.1"
-          }
-        }
-        browsers = YAML.safe_load(File.read("#{__dir__}/browsers_bs.yml"))
-        capabilities.deep_merge! browsers[browser_type]
-        capabilities.deep_merge! JSON.parse(capabilities_option)
-        Selenium::WebDriver::Remote::Capabilities.new capabilities
-      end
-
       # @param device_type [String]
       def for_bitbar_device(bitbar_api_key, device_type, platform, platform_version, capabilities_option)
         capabilities = {