+++ lib/maze/hooks/browser_hooks.rb
@@ -25,12 +25,12 @@ module Maze
         when :bs
           # BrowserStack browser
           tunnel_id = SecureRandom.uuid
-          config.capabilities = Maze::Capabilities.for_browser_stack_browser config.browser,
-                                                                             tunnel_id,
-                                                                             config.capabilities_option
-          Maze::BrowserStackUtils.start_local_tunnel config.bs_local,
-                                                     tunnel_id,
-                                                     config.access_key
+          config.capabilities = Maze::Farm::BrowserStack::Capabilities.browser config.browser,
+                                                                               tunnel_id,
+                                                                               config.capabilities_option
+          Maze::Farm::BrowserStack::Utils.start_local_tunnel config.bs_local,
+                                                             tunnel_id,
+                                                             config.access_key
         end
 
         # Create and start the relevant driver
@@ -53,7 +53,7 @@ module Maze
 
       def at_exit
         if Maze.config.farm == :bs
-          Maze::BrowserStackUtils.stop_local_tunnel
+          Maze::Farm::BrowserStack::Utils.stop_local_tunnel
           Maze.driver.driver_quit
         end
       end